backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
spleeter==2.4.0
ffmpeg-python==0.2.0
python-dotenv==1.0.1
